+++ pdf/lib/src/page_format.dart
@@ -66,7 +66,17 @@ class PdfPageFormat {
         marginRight: marginRight ?? this.marginRight);
   }
 
+  /// Total page dimensions
   PdfPoint get dimension => PdfPoint(width, height);
+
+  /// Total page width excluding margins
+  double get availableWidth => width - marginLeft - marginRight;
+
+  /// Total page height excluding margins
+  double get availableHeight => height - marginTop - marginBottom;
+
+  /// Total page dimensions excluding margins
+  PdfPoint get availableDimension => PdfPoint(availableWidth, availableHeight);
 
   PdfPageFormat get landscape =>
       width >= height ? this : copyWith(width: height, height: width);
@@ -85,6 +95,6 @@ class PdfPageFormat {
 
   @override
   String toString() {
-    return '${width}x$height';
+    return 'Page ${width}x$height margins:$marginLeft, $marginTop, $marginRight, $marginBottom';
   }
 }